refactor: Enhance MCP module with async configuration support and new core module
